Reorganise the "base" object
