+++ src/base.ts
@@ -1,0 +1,5 @@
+import {delay} from "./delay";
+import * as maths from "./maths";
+import * as user from "./user";
+
+export {delay, maths, user};
