+++ src/runner.ts
@@ -1,13 +1,11 @@
 import * as dopri from "dopri";
-import type { UserType } from "./user";
-import * as userHelpers from "./user";
 
-// Probably this is something that dopri should export for us, we
-// could also use its types for the rhs and output members below.
-type Solution = (t: number) => number[];
+import * as base from "./base";
+import type { Solution } from "./delay";
+import type { UserType } from "./user";
 
 export type OdinModelConstructable =
-    new(userHelpers: any, pars: UserType, unknownAction: string) => OdinModel;
+    new(base: any, pars: UserType, unknownAction: string) => OdinModel;
 
 interface OdinModelODE {
     initial(t: number): number[];
@@ -28,21 +26,6 @@ function isDDEModel(model: OdinModel): model is OdinModelDDE {
 }
 
 export type OdinModel = OdinModelODE | OdinModelDDE;
-
-export function delay(solution: Solution, t: number, index: number[],
-                      state: number[]) {
-    // Later, we'll update dopri.js to allow passing index here,
-    // which will make this more efficient. However, no change to
-    // the external interface will be neeed.
-    const y = solution(t);
-    for (let i = 0; i < index.length; ++i) {
-        state[i] = y[index[i]];
-    }
-}
-
-export const base = {checkUser: userHelpers.checkUser,
-                     delay,
-                     setUserScalar: userHelpers.setUserScalar};
 
 // tslint:disable-next-line:variable-name
 export function wodinRun(Model: OdinModelConstructable, pars: UserType,
